+++ auctions/feature_iterators.Template.h
@@ -157,7 +157,7 @@ CrossProductIterator<Pred>::number_remaining () const
   if (! valid() ) return 0;
   int n (0);
   for (unsigned iSlow=0; iSlow < mSlowSource.size(); ++iSlow)
-    n += mFastSource.size() - mFastIndices[iSlow];
+    n += (int)mFastSource.size() - mFastIndices[iSlow];
   debugging::debug("CPIT",4) << " CP of vectors of sizes " << mSlowSource.size() << " x " << mFastSource.size()
 			     << " has " << n << " remaining features." << std::endl;
   return n;
